Removes "forgot password" from welcome page
This button  wasn't displaying the popup message about contacting
ePRaSE administrators, as does happen correctly when you click the
same button from the login page. I started out trying to make the
popup happen from the welcome page when it occurred to me the
'forgot password' link doesn't even really need to be on the
welcome page. It belongs on the login page.

So the easier fix was just to remove it from welcome. It's just
commented out if anyone decides they really want it back in
there.