Changes "System complete" to "System information completed"
Per NHS request, when the assessment state shows "System
complete" on the list of partially completed assessments, it
might be misinterpreted to mean the whole assessment is
finished. So this commit changes that text to "System
information completed" while leaving the underlying stuff that
gets stored in the database alone. (Changing only that display,
not the actual state options.)